opensand-testbed/requirements.txt
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
docker==4.1.0
idna==2.10
loguru==0.3.2
nclib==0.8.3
numpy==1.20.0
pypiwin32==223
python-dotenv==0.15.0
pywin32==300
requests==2.25.1
six==1.15.0
urllib3==1.26.3
websocket-client==0.57.0
win32-setctime==1.0.3
